+++ scout/server/blueprints/panels/forms.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from flask_wtf import FlaskForm
-from wtforms import (BooleanField, TextField, SelectMultipleField)
+from wtforms import (BooleanField, TextField, SelectMultipleField, FloatField)
 
 from scout.constants import GENETIC_MODELS
 
@@ -9,5 +9,6 @@ class PanelGeneForm(FlaskForm):
     disease_associated_transcripts = SelectMultipleField('Disease transcripts', choices=[])
     reduced_penetrance = BooleanField()
     mosaicism = BooleanField()
+    entry_version = FloatField()
     inheritance_models = SelectMultipleField(choices=GENETIC_MODELS)
     comment = TextField()